Hw5 auth

+++ src/controllers/contacts.js
@@ -18,6 +18,7 @@ export async function getContactsController(req, res) {
     perPage,
     sortBy,
     sortOrder,
+    userId: req.user._id,
   });
 
   res.status(200).json({
@@ -35,6 +36,10 @@ export async function getContactController(req, res) {
     throw createHttpError(404, 'Contact not found');
   }
 
+  if (contact.userId.toString() !== req.user._id.toString()) {
+    throw createHttpError(404, 'Contact not found');
+  }
+
   res.status(200).json({
     status: 200,
     message: `Successfully found contact with id ${contactId}!`,
@@ -43,22 +48,33 @@ export async function getContactController(req, res) {
 }
 
 export async function createContactController(req, res) {
-  const contact = await createContact(req.body);
+  const contact = {
+    ...req.body,
+    userId: req.user._id,
+  };
+
+  const result = await createContact(contact);
 
   res.status(201).json({
     status: 201,
     message: 'Successfully created a contact!',
-    data: contact,
+    data: result,
   });
 }
 
 export async function updateContactController(req, res, next) {
   const { contactId } = req.params;
-  const result = await updateContact(contactId, req.body);
+  const contact = await getContactById(contactId);
 
-  if (!result) {
+  if (!contact) {
     throw createHttpError(404, 'Contact not found');
   }
+
+  if (contact.userId.toString() !== req.user._id.toString()) {
+    throw createHttpError(404, 'Contact not found');
+  }
+
+  const result = await updateContact(contactId, req.body);
 
   res.status(200).json({
     status: 200,
@@ -69,12 +85,17 @@ export async function updateContactController(req, res, next) {
 
 export async function deleteContactController(req, res) {
   const { contactId } = req.params;
-
-  const contact = await deleteContact(contactId);
+  const contact = await getContactById(contactId);
 
   if (!contact) {
     throw createHttpError(404, 'Contact not found');
   }
+
+  if (contact.userId.toString() !== req.user._id.toString()) {
+    throw createHttpError(404, 'Contact not found');
+  }
+
+  await deleteContact(contactId);
 
   res.status(204).send();
 }
